Baseline_models(MGAT)/requirements.txt
matplotlib==3.7.0
numpy==1.24.2
torch_geometric==2.2.0
torchcontrib==0.0.2

